pokemon/requirements.txt
beautifulsoup4==4.12.3
Flask==3.0.1
pandas==1.4.3
pymongo==4.6.1
python-dotenv==1.0.1
Requests==2.31.0
